pyproject.toml
[project]
name = "opencuff"
version = "0.1.0"
description = "Controlled and secure operations for coding agents"
readme = "README.md"
authors = [
    { name = "Kfir Gollan" }
]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.14.3",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "watchfiles>=0.21",
    "structlog>=24.0",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.13",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
